Run `yarn run build` and make required mods
The package.json modifications were added automatically by the script.
The .gitignore was not. I don't know yet whether I'll stick with a
`build` folder or whether a `lib` one will be better. Since this is a
react project it makes sense to stick with the react standard `build` I
suppose.