Favour not using the container approach
testing-library prefers that you approach identifying your components in
the screen as a human would. I think this is more applicable to tests
involving user interactions, but I'll try and be idiomatic. Using
`firstChild` seems only as fragile as what I would have to resort to if
the styling that I'm checking wasn't actually *on* the div with the
styling.