Favour single style test
I have had to remove the rgba color definition since it appears that
something here doesn't like it. It gets removed from the component style
as well as the test just like any style would if it were invalid. If we
wrote bluue rather than blue for example. It is just not included in the
component style or the test style which can lead to false positives.
Need to be careful of this.