Add undefined hole interaction rendering
It needs to basically be invisible, but we have to present a '/' so that
the grid layout won't collapse in the absense of any values. This is a
property of grid layouts which I haven't researched to determine whether
it's configurable. While I don't know and / or while it's not, making
the content invisible will do.