Remove src/index.ts as it caused total failure
It's unclear to me what happened here. The tests and linting passed, but
the running app just gave the impression that javascript was turned off,
based on the fact that the <noscript> tag becomes active. I can imagine
that the problem is that there are two files called index in the same
folder, just with different extensions (ts & tsx).

REVISIT: this tag here so that I can come back to this commit in the
future.