Move theme within HarpFace
I believe there might be at least 2 different component packages being
created here. The one to display the harp and the other for the UI.

If that's true then it would make sense for each package to own it's
theme but to be informed by the composing component if required.