Add theme colours with defaults for overriding
The seed size will default to 1 now as well in a similar fashion. Spot
the difference in how the default was set before this change. The
seedSize was a mandatory part of the primer type, but it no longer is.

I generally disapprove of optional types, but for this commit at least
it is convenient. In an upcoming commit I plan to make them required
again.