Fix linting errors caused by new interface requirements
It appears that PitchIds isn't actually exported from harpstrata which
means that there are still linting errors now, but there shouldn't be if
harpstrata is exporting everything that it should. Perhaps requires a
patch version bump to harpstrata, but it might also just be an ide
error. This is the next thing for me to investigate. If you see no more
commits referencing it then it is likely that it was just an ide error.