Add HarpGuru test for pitch altering functionality
I had to make quite a few changes lower down here because the harpstrata
version which we are currently using only makes C & Db pitches
available. Errors were occuring when running the tests with A & Bb as we
already had in the PitchButtons component, presumably because an error
was occurring in harpstrata trying to handle a value which it doesn't
have a mapping for. I have raised an issue for errors to be handled
deliberately in harpstrata -
https://github.com/js-jslog/harpstrata/issues/12