Add local state for activePozition
Really I would rather just have state for the harpstrata and deduce
these other states from it, but this is quicker and ultimately I think
the harpstrata itself should return the parameters which were used to
generate it. This may be included in an upcoming version. For the
purposes of this pull request, I'll just stick with the additional state
objects.