Add test and functionality for hole stasis
When a new Pozition is selected for example, it may be the
activePitchIds or the activeDegreeIds which are updated, but not both.
Making the decision of which one to update be governed by which display
mode is active means that we are giving priority to the Ids which the
user is currently looking at, allowing them to change the harp while
keeping the pitches or degrees which they have selected.